chore(cli): move sinks into their own folder
