+++ packages/aws-cdk/lib/cli/telemetry/sinks/endpoint-sink.ts
@@ -3,10 +3,10 @@ import type { Agent } from 'https';
 import { request } from 'https';
 import { parse, type UrlWithStringQuery } from 'url';
 import { ToolkitError } from '@aws-cdk/toolkit-lib';
-import { IoHelper } from '../../api-private';
-import type { IIoHost } from '../io-host';
-import type { TelemetrySchema } from './schema';
-import type { ITelemetrySink } from './sink-interface';
+import { IoHelper } from '../../../api-private';
+import type { IIoHost } from '../../io-host';
+import type { TelemetrySchema } from '../schema';
+import type { ITelemetrySink } from '../sink-interface';
 
 const REQUEST_ATTEMPT_TIMEOUT_MS = 500;
