+++ packages/aws-cdk/lib/cli/telemetry/sinks/io-host-sink.ts
@@ -1,7 +1,7 @@
 import type { IIoHost } from '@aws-cdk/toolkit-lib';
-import type { TelemetrySchema } from './schema';
-import type { ITelemetrySink } from './sink-interface';
-import { IoHelper } from '../../api-private';
+import type { TelemetrySchema } from '../schema';
+import type { ITelemetrySink } from '../sink-interface';
+import { IoHelper } from '../../../api-private';
 
 /**
  * Properties for the StdoutTelemetryClient
