+++ packages/aws-cdk/test/cli/telemetry/sink/endpoint-sink.test.ts
@@ -1,8 +1,8 @@
 import * as https from 'https';
-import { IoHelper } from '../../../lib/api-private';
-import { CliIoHost } from '../../../lib/cli/io-host';
-import { EndpointTelemetrySink } from '../../../lib/cli/telemetry/endpoint-sink';
-import type { EventType, TelemetrySchema } from '../../../lib/cli/telemetry/schema';
+import { IoHelper } from '../../../../lib/api-private';
+import { CliIoHost } from '../../../../lib/cli/io-host';
+import { EndpointTelemetrySink } from '../../../../lib/cli/telemetry/endpoint-sink';
+import type { EventType, TelemetrySchema } from '../../../../lib/cli/telemetry/schema';
 
 // Mock the https module
 jest.mock('https', () => ({
